Introduce the concept of a label
