+++ text/src/lib.rs
@@ -1,5 +1,9 @@
 mod distance_field_gen;
+mod label;
 mod point;
 mod render_graph;
+mod vertex;
 
+pub use label::*;
 pub use render_graph::render_graph;
+pub use vertex::*;
