+++ text/src/vertex.rs
@@ -1,0 +1,52 @@
+//! TODO: This belongs to shell.
+
+use std::mem;
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+pub struct Vertex {
+    position: [f32; 3],
+}
+
+impl Vertex {
+    fn new(x: f32, y: f32, z: f32) -> Self {
+        Self {
+            position: [x, y, z],
+        }
+    }
+
+    fn desc() -> &'static wgpu::VertexBufferLayout<'static> {
+        const LAYOUT: wgpu::VertexBufferLayout = wgpu::VertexBufferLayout {
+            array_stride: mem::size_of::<Vertex>() as wgpu::BufferAddress,
+            step_mode: wgpu::VertexStepMode::Vertex,
+            attributes: &wgpu::vertex_attr_array![0 => Float32x3],
+        };
+
+        &LAYOUT
+    }
+}
+
+impl From<(f32, f32, f32)> for Vertex {
+    fn from(v: (f32, f32, f32)) -> Self {
+        Self::new(v.0, v.1, v.2)
+    }
+}
+
+#[repr(C)]
+#[derive(Copy, Clone, Debug, bytemuck::Pod, bytemuck::Zeroable)]
+pub struct TextureVertex {
+    pub position: Vertex,
+    pub tex_coords: [f32; 2],
+}
+
+impl TextureVertex {
+    pub fn desc() -> &'static wgpu::VertexBufferLayout<'static> {
+        const LAYOUT: wgpu::VertexBufferLayout = wgpu::VertexBufferLayout {
+            array_stride: mem::size_of::<TextureVertex>() as wgpu::BufferAddress,
+            step_mode: wgpu::VertexStepMode::Vertex,
+            attributes: &wgpu::vertex_attr_array![0 => Float32x3, 1 => Float32x2],
+        };
+
+        &LAYOUT
+    }
+}
